+++ @ornikar/jest-config/test-setup-after-env.js
@@ -18,6 +18,10 @@ failOnConsole({
   silenceMessage: (message) => {
     // Do not ignore the act / await warning anymore once we are on React 18
     if (message.includes('You called act(async () => ...) without await')) return true;
+
+    // Warning from @react-aria/ssr
+    if (message.startsWith('In React 18, SSRProvider is not necessary')) return true;
+
     return deprecatedReactLifeCycleMethods.some((lifecycleMethod) =>
       message.includes(`${lifecycleMethod} has been renamed, and is not recommended for use.`),
     );