+++ @ornikar/repo-config-react-legacy-css/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.0.7](https://github.com/ornikar/shared-configs/compare/@ornikar/repo-config-react-legacy-css@11.0.6...@ornikar/repo-config-react-legacy-css@11.0.7) (2023-07-25)
+
+**Note:** Version bump only for package @ornikar/repo-config-react-legacy-css
+
+
+
+
+
+## [11.0.6](https://github.com/ornikar/shared-configs/compare/@ornikar/repo-config-react-legacy-css@11.0.5...@ornikar/repo-config-react-legacy-css@11.0.6) (2023-07-19)
+
+**Note:** Version bump only for package @ornikar/repo-config-react-legacy-css
+
+
+
+
+
+## [11.0.5](https://github.com/ornikar/shared-configs/compare/@ornikar/repo-config-react-legacy-css@11.0.4...@ornikar/repo-config-react-legacy-css@11.0.5) (2023-07-13)
+
+**Note:** Version bump only for package @ornikar/repo-config-react-legacy-css
+
+
+
+
+
+## [11.0.4](https://github.com/ornikar/shared-configs/compare/@ornikar/repo-config-react-legacy-css@11.0.3...@ornikar/repo-config-react-legacy-css@11.0.4) (2023-07-05)
+
+**Note:** Version bump only for package @ornikar/repo-config-react-legacy-css
+
+
+
+
+
 ## [11.0.3](https://github.com/ornikar/shared-configs/compare/@ornikar/repo-config-react-legacy-css@11.0.2...@ornikar/repo-config-react-legacy-css@11.0.3) (2023-05-23)
 
 **Note:** Version bump only for package @ornikar/repo-config-react-legacy-css