+++ @ornikar/storybook-config/CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.1.1](https://github.com/ornikar/shared-configs/compare/@ornikar/storybook-config@7.1.0...@ornikar/storybook-config@7.1.1) (2023-07-19)
+
+**Note:** Version bump only for package @ornikar/storybook-config
+
+
+
+
+
+## [7.1.0](https://github.com/ornikar/shared-configs/compare/@ornikar/storybook-config@7.0.3...@ornikar/storybook-config@7.1.0) (2023-06-26)
+
+
+### Features
+
+* **webpack-config:** export workspaceAliases ([2eba7bb](https://github.com/ornikar/shared-configs/commit/2eba7bbe22372d76d1ba87a7172f98315cbbff7b))
+
+
+
 ## [7.0.3](https://github.com/ornikar/shared-configs/compare/@ornikar/storybook-config@7.0.2...@ornikar/storybook-config@7.0.3) (2023-05-05)
 
 **Note:** Version bump only for package @ornikar/storybook-config